v30.2.1 Update

+++ updateFiles/wombat_update.sh
@@ -1,15 +1,45 @@
 #!/bin/bash
 
-FW_VERSION=30.2.0
+HOME=/home/kipr
+FW_VERSION=$(cat "$HOME/wombat-os/configFiles/board_fw_version.txt")
 
 echo "   "
 echo "Starting Wombat Update #$FW_VERSION"
 echo "..."
 
-HOME=/home/kipr
+###############################
+#
+# Check for Current FW Version
+#
+###############################
+
+# Check if board_fw_version.txt exists
+if [ ! -f /usr/share/kipr/board_fw_version.txt ]; then
+    echo "Your version is too old to update. Please reflash your SD card."
+    exit 1
+fi
+
+###############################
+#
+# Move update files
+#
+###############################
+
+# Change to updateFiles directory and copy updateMe.sh to home directory
 cd $HOME/wombat-os/updateFiles
 cp files/updateMe.sh $HOME
 sudo chmod u+x $HOME/updateMe.sh
+
+# Change to configFiles directory and copy board_fw_version.txt to kipr share directory
+cd $HOME/wombat-os/configFiles
+if [ ! -d /usr/share/kipr ]; then
+    sudo mkdir /usr/share/kipr
+fi
+
+sudo scp board_fw_version.txt /usr/share/kipr/
+sudo scp create3_server.service /etc/systemd/system/
+sudo scp journal.conf /etc/systemd/journald.conf
+
 
 ###############################
 #
@@ -18,7 +48,7 @@ sudo chmod u+x $HOME/updateMe.sh
 ###############################
 
 #remount root filesystem as read write
-mount -o remount,rw /
+# mount -o remount,rw /
 
 
 ###############################
@@ -32,7 +62,7 @@ echo "Updating harrogate..."
 sudo rm -r /home/kipr/harrogate
 sudo tar -C /home/kipr -zxvf pkgs/harrogate.tar.gz
 sudo chmod -R 777 /home/kipr/harrogate
-cd /home/kipr/harrogate
+cd $HOME/harrogate
 echo "Installing harrogate dependencies..."
 npm install browserfy
 npm install
